examples/example-ai-langgraph/pyproject.toml
[project]
name = "example-ai-langgraph"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.3.0",
    "opentelemetry-instrumentation-langchain>=0.24.0",
    "opentelemetry-sdk>=1.30.0",
    "opentelemetry-exporter-otlp-proto-http>=1.30.0",
]
